feat(#56): Implemented Console Output
Added an attribute of class ConsoleOutput in Simulator class. You can use the method setMessage(String) to add to the console output. For example if we take the object of SimulatorMap(0), we can do the following in the controller:

Simulator sim = simulatorMap.get(0);
sim.getcOut().setMessage("test");

And then in the JS file, add the following at the end of your function

displayConsoleOut();